% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vir.R
\name{create_tbl_pathway2kID}
\alias{create_tbl_pathway2kID}
\title{Create a table of KEGG pathways and their associated KO numbers}
\usage{
create_tbl_pathway2kID(pathway_names)
}
\arguments{
\item{pathway_names}{A character vector of KEGG pathway names.}
}
\value{
A data frame with columns:
  \item{pathway_name}{KEGG pathway name}
  \item{pathway_id}{KEGG pathway ID}
  \item{kID}{KEGG Orthology (KO) number}
  \item{description}{KEGG Orthology description}
}
\description{
This function retrieves the KO numbers and descriptions for a list of KEGG pathways.
}
