% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vir.R
\name{parse_aniclst}
\alias{parse_aniclst}
\title{Parse AniCLST Output (Clusters based on Average Nucleic Acid Identity)}
\usage{
parse_aniclst(fin_aniclst)
}
\arguments{
\item{fin_aniclst}{A character string specifying the path to the AniCLST
output file.}
}
\value{
A data frame with columns `contig_repid`, `contig_id` and `vOTU_id`,
  which maps contig IDs to representative contig IDs and unique vOTU IDs for
  easy identification of clusters.
}
\description{
This function parses the AniCLST output, which maps contigs to their
corresponding representative contig IDs (Rep IDs). The function assigns
unique vOTU IDs to each Rep ID and creates a data frame mapping each
contig ID to a vOTU ID.
}
\details{
The function performs the following steps:
\enumerate{
  \item Reads the AniCLST output file.
  \item Renames columns for consistent naming conventions.
  \item Generates a unique `vOTU_id` for each unique Rep ID.
  \item Maps `contig_id` values to `vOTU_id` values for downstream use.
}
}
\examples{
\dontrun{
# Example usage:
# Assuming 'aniclst_output.txt' is the AniCLST output file
df_ctg2vOTU <- parse_aniclst("aniclst_output.txt")
}
}
