% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vir.R
\name{parse_bacphlip}
\alias{parse_bacphlip}
\title{Parse BACPHLIP Output}
\usage{
parse_bacphlip(fin_bacphlip)
}
\arguments{
\item{fin_bacphlip}{Character string specifying the path to the
BACPHLIP output file.}
}
\value{
A data frame containing contig IDs and predicted
  lifestyles as `bacphlip_lifestyle`.
}
\description{
This function parses the output of BACPHLIP software, which
predicts the lifestyle (temperate or virulent) of phage genomes.
}
\examples{
\dontrun{
# Assuming 'bacphlip_output.txt' is the BACPHLIP output file
df_bacphlip <- parse_bacphlip("bacphlip_output.txt")
}
}
