% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vir.R
\name{read_coverm_vir}
\alias{read_coverm_vir}
\title{Read CoverM output file}
\usage{
read_coverm_vir(fpath, fbin2contig = 0, cov = 0)
}
\arguments{
\item{fpath}{CoverM output file name}

\item{fbin2contig}{Contig binning to contig mapping}

\item{cov}{Coverage fraction file or not}
}
\value{
A dataframe
}
\description{
Read CoverM output file
}
